library_frontend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
pika==1.3.2
sqlparse==0.5.1
